requirements.txt
beautifulsoup4==4.12.3
boto3==1.34.69
botocore==1.34.69
discord.py==2.3.2
python-dotenv==1.0.1
Requests==2.31.0
setuptools==69.0.2
